An S4 class representing the evaluation of a metric run by simulator.

An object of class Evals consists of information to identify the model, draws, method, and metric objects this output was derived from. It also has a list called evals, which is where the output of the metric is stored. Currently, the labels of all these objects are also included so that plot functions can use human-readable labels without requiring re-loading these.

Slots

model_name

the name of the Model object this output is derived from.

model_label

the label of the Model object this output is derived from.

index

the index of the Draws object this output is derived from.

method_name

the name of the Method object this output is derived from.

method_label

the label of the Method object this output is derived from.

metric_name

the name of the Metric object this output is derived from.

metric_label

the label of the Metric object this output is derived from.

evals

a named list with each element labeled by a method_name each evals[[m]] is itself a named list with each element labeled as ri.j where i is the index and j ranges from 1 to nsim. Element out$ri.j is output of metric metric_name on random draw ri.j.

See also